+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.0
+* Add firebase analytics route filter function
+* Add feature of hide dialog at global scope
+
 ## 1.0.1
 Fix android back button last view stack deletion.